requirements.txt
pandas
pytest
streamlit
google-generativeai
python-dotenv
requests
